HIPFORGE-384 Move review nodetype to common module
